Move former pages to client components
