+++ app/add/page.tsx
@@ -1,0 +1,5 @@
+import AddPage from "../../src/pages/AddPage";
+
+export default async function Add() {
+  return <AddPage />
+};
