--- pages/_app.tsx
@@ -1,10 +1,0 @@
-import type { AppProps } from 'next/app'
-import Layout from '../src/components/Layout'; 
-
-export default function App({ Component, pageProps }: AppProps) {
-  return (
-    <Layout>
-      <Component {...pageProps} />
-    </Layout>
-  );
-}
